Added Player Category, Moved Some of The Modules into Player Category
