+++ src/main/java/me/xakeplusplus/spg/module/Category.java
@@ -2,7 +2,7 @@ package me.xakeplusplus.spg.module;
 
 public enum Category {
     COMBAT("COMBAT"), MOVEMENT("MOVEMENT"), MISC("MISC"), EXPLOITS("EXPLOITS"),
-    CHAT("CHAT"), RENDER("RENDER");
+    CHAT("CHAT"), PLAYER("PLAYER"), RENDER("RENDER");
 
     public String name;
     public int moduleIndex;
